+++ packages/app/components/login/login-with-apple.tsx
@@ -1,11 +1,7 @@
-import { createWalletClient, custom } from "viem";
-import { mainnet } from "viem/chains";
-
 import { useRouter } from "@showtime-xyz/universal.router";
 
 import { useAuth } from "app/hooks/auth/use-auth";
 import { LOGIN_MAGIC_ENDPOINT } from "app/hooks/auth/use-magic-login";
-import { useWeb3 } from "app/hooks/use-web3";
 import { Logger } from "app/lib/logger";
 import { useMagic } from "app/lib/magic";
 import { useMagicSocialAuth } from "app/lib/social-logins";
@@ -17,8 +13,6 @@ export const LoginWithApple = () => {
   const { setAuthenticationStatus, login, logout } = useAuth();
   const router = useRouter();
   const { performMagicAuthWithApple } = useMagicSocialAuth();
-  const { setWeb3 } = useWeb3();
-  const { magic } = useMagic();
 
   return (
     <LoginButton
@@ -33,13 +27,6 @@ export const LoginWithApple = () => {
             provider_access_token: result.oauth.accessToken,
             provider_scope: result.oauth.scope,
           });
-
-          const client = createWalletClient({
-            chain: mainnet,
-            transport: custom(magic.rpcProvider),
-          });
-
-          setWeb3({ ...client, isMagic: true });
 
           // when profile is incomplete, login will automatically redirect user to /profile/edit. So we don't need to redirect user to decodedURI
           if (!isProfileIncomplete(user.data.profile)) {
