+++ packages/app/components/login/login-with-google.tsx
@@ -5,7 +5,6 @@ import { useRouter } from "@showtime-xyz/universal.router";
 
 import { useAuth } from "app/hooks/auth/use-auth";
 import { LOGIN_MAGIC_ENDPOINT } from "app/hooks/auth/use-magic-login";
-import { useWeb3 } from "app/hooks/use-web3";
 import { Logger } from "app/lib/logger";
 import { useMagic } from "app/lib/magic";
 import { useMagicSocialAuth } from "app/lib/social-logins";
@@ -17,7 +16,6 @@ export const LoginWithGoogle = () => {
   const { setAuthenticationStatus, login, logout } = useAuth();
   const router = useRouter();
   const { performMagicAuthWithGoogle } = useMagicSocialAuth();
-  const { setWeb3 } = useWeb3();
   const { magic } = useMagic();
 
   return (
@@ -34,12 +32,6 @@ export const LoginWithGoogle = () => {
             provider_scope: result.oauth.scope,
           });
 
-          const client = createWalletClient({
-            chain: mainnet,
-            transport: custom(magic.rpcProvider),
-          });
-
-          setWeb3({ ...client, isMagic: true });
           // when profile is incomplete, login will automatically redirect user to /profile/edit. So we don't need to redirect user to decodedURI
           if (!isProfileIncomplete(user.data.profile)) {
             router.pop();
