+++ packages/app/components/settings/settings-email-item.tsx
@@ -1,4 +1,4 @@
-import { useCallback, useEffect, useState } from "react";
+import { useState } from "react";
 
 import { DataPill } from "@showtime-xyz/universal.data-pill";
 import { CheckFilled1 } from "@showtime-xyz/universal.icon";
@@ -6,10 +6,7 @@ import { colors } from "@showtime-xyz/universal.tailwind";
 import { Text } from "@showtime-xyz/universal.text";
 import { View } from "@showtime-xyz/universal.view";
 
-import { useCurrentUserAddress } from "app/hooks/use-current-user-address";
 import { useManageAccount } from "app/hooks/use-manage-account";
-import { useWeb3 } from "app/hooks/use-web3";
-import { useMagic } from "app/lib/magic";
 import { WalletAddressesV2 } from "app/types";
 
 import { DropdownMenu } from "./dropdown-menu";
@@ -25,32 +22,10 @@ export type EmailHeaderProps = {
 };
 
 export const SettingsEmailItem = (props: EmailItemProps) => {
-  const [isCurrentEmail, setIsCurrentEmail] = useState(false);
+  const [isCurrentEmail] = useState(false);
   const { removeAccount } = useManageAccount();
-  const { magic } = useMagic();
-  const { isMagic } = useWeb3();
-  const { userAddress } = useCurrentUserAddress();
   const email = props.email;
   const backendAddress = props.address;
-
-  const getCurrentMagicUser = useCallback(async () => {
-    if (isMagic) {
-      const magicMetaData = await magic?.user?.getMetadata();
-      const currentEmail = magicMetaData.email;
-      const currentMagicAddress = magicMetaData.publicAddress;
-      const isMatchingMagicEmail =
-        currentEmail?.toLowerCase() === email?.toLowerCase();
-      const isMatchingMagicAddress =
-        currentMagicAddress?.toLowerCase() === userAddress?.toLowerCase();
-      if (isMatchingMagicEmail && isMatchingMagicAddress) {
-        setIsCurrentEmail(true);
-      }
-    }
-  }, [magic, isMagic, email, userAddress]);
-
-  useEffect(() => {
-    getCurrentMagicUser();
-  }, [getCurrentMagicUser]);
 
   return (
     <View tw="w-full flex-row items-center justify-between px-4 py-5 lg:px-0">
