+++ packages/app/hooks/auth/use-magic-login.ts
@@ -3,7 +3,6 @@ import { useCallback } from "react";
 import { createWalletClient, custom } from "viem";
 import { mainnet } from "viem/chains";
 
-import { useWeb3 } from "app/hooks/use-web3";
 import {
   BYPASS_EMAIL,
   BYPASS_EMAIL_WITH_INSECURE_KEYS,
@@ -18,7 +17,6 @@ export const LOGIN_MAGIC_ENDPOINT = "login_magic";
 export function useMagicLogin() {
   //#region hooks
   const { setAuthenticationStatus, login, logout } = useAuth();
-  const { setWeb3 } = useWeb3();
   const { magic, Magic } = useMagic();
   //#endregion
 
@@ -35,19 +33,12 @@ export function useMagicLogin() {
           did,
           phone_number: phoneNumber,
         });
-
-        const client = createWalletClient({
-          chain: mainnet,
-          transport: custom(magic.rpcProvider),
-        });
-
-        setWeb3({ ...client, isMagic: true });
       } catch (error) {
         logout();
         throw error;
       }
     },
-    [magic, login, logout, setAuthenticationStatus, setWeb3]
+    [magic, login, logout, setAuthenticationStatus]
   );
 
   const loginWithEmail = useCallback(
@@ -89,19 +80,12 @@ export function useMagicLogin() {
           did,
           email: overrideEmail,
         });
-
-        const client = createWalletClient({
-          chain: mainnet,
-          transport: custom(magic.rpcProvider),
-        });
-
-        setWeb3({ ...client, isMagic: true });
       } catch (error) {
         logout();
         throw error;
       }
     },
-    [magic, Magic, login, logout, setAuthenticationStatus, setWeb3]
+    [magic, Magic, login, logout, setAuthenticationStatus]
   );
 
   //#endregion
