+++ packages/app/hooks/use-wallet/use-wallet.ts
@@ -1,13 +1,13 @@
-import { useState, useEffect, useMemo, useRef } from "react";
+import { useEffect, useMemo, useRef } from "react";
 
 import { createWalletClient, custom } from "viem";
 import { mainnet } from "viem/chains";
 
 import { useLatestValueRef } from "app/hooks/use-latest-value-ref";
 import { useWalletMobileSDK } from "app/hooks/use-wallet-mobile-sdk";
-import { useWeb3 } from "app/hooks/use-web3";
 import { useWeb3Modal } from "app/lib/react-native-web3-modal";
 
+import { useStableCallback } from "../use-stable-callback";
 import { ConnectResult, UseWalletReturnType } from "./types";
 import { useRandomWallet } from "./use-random-wallet";
 
@@ -15,31 +15,12 @@ const useWallet = (): UseWalletReturnType => {
   const walletConnectedPromiseResolveCallback = useRef<any>(null);
   const walletDisconnectedPromiseResolveCallback = useRef<any>(null);
   const web3Modal = useWeb3Modal();
-  const { web3, isMagic, magicWalletAddress, getWalletClient } = useWeb3();
   const mobileSDK = useWalletMobileSDK();
-  const [address, setAddress] = useState<`0x${string}` | undefined>();
-
   // we use this hook to prevent stale values in closures
   const walletConnectInstanceRef = useLatestValueRef(web3Modal);
   const coinbaseMobileSDKInstanceRef = useLatestValueRef(mobileSDK);
 
   const walletConnected = web3Modal.isConnected || mobileSDK.connected;
-  useEffect(() => {
-    (async function fetchUserAddress() {
-      if (web3Modal.address) {
-        setAddress(web3Modal.address as `0x${string}`);
-      } else if (magicWalletAddress) {
-        setAddress(magicWalletAddress as `0x${string}`);
-      } else if (mobileSDK.address) {
-        setAddress(mobileSDK.address as `0x${string}`);
-      } else if (web3) {
-        const address = web3.account?.address;
-        setAddress(address);
-      } else {
-        setAddress(undefined);
-      }
-    })();
-  }, [web3, magicWalletAddress, mobileSDK.address, web3Modal.address]);
 
   // WalletConnect connected
   useEffect(() => {
@@ -92,6 +73,33 @@ const useWallet = (): UseWalletReturnType => {
     }
   }, [mobileSDK]);
 
+  const getWalletClient = useStableCallback(async () => {
+    if (web3Modal.isConnected && web3Modal.provider) {
+      const client = createWalletClient({
+        chain: mainnet,
+        transport: custom(web3Modal.provider),
+      });
+      return client;
+    } else if (mobileSDK.connected && mobileSDK.address) {
+      const MobileSDKProvider = (
+        await import(
+          "@coinbase/wallet-mobile-sdk/build/WalletMobileSDKEVMProvider"
+        )
+      ).WalletMobileSDKEVMProvider;
+      const mobileSDKProvider = new MobileSDKProvider({
+        jsonRpcUrl: `https://mainnet.infura.io/v3/${process.env.NEXT_PUBLIC_INFURA_ID}`,
+        address: mobileSDK.address,
+      });
+
+      const client = createWalletClient({
+        chain: mainnet,
+        transport: custom(mobileSDKProvider as any),
+      });
+
+      return client;
+    }
+  });
+
   const result = useMemo(() => {
     const wcConnected = web3Modal.isConnected;
     const mobileSDKConnected = mobileSDK.connected;
@@ -104,7 +112,6 @@ const useWallet = (): UseWalletReturnType => {
     }
 
     return {
-      address,
       connect: async () => {
         walletConnectInstanceRef.current.open();
         return new Promise<ConnectResult>((resolve) => {
@@ -114,7 +121,8 @@ const useWallet = (): UseWalletReturnType => {
       disconnect: async () => {
         if (walletConnectInstanceRef.current.isConnected) {
           walletConnectInstanceRef.current.disconnect();
-        } else if (coinbaseMobileSDKInstanceRef.current.connected) {
+        }
+        if (coinbaseMobileSDKInstanceRef.current.connected) {
           coinbaseMobileSDKInstanceRef.current.disconnect();
         }
 
@@ -123,9 +131,7 @@ const useWallet = (): UseWalletReturnType => {
         });
       },
       name: walletName,
-      connected: walletConnected || isMagic,
-      isMagicWallet: isMagic,
-      walletClient: web3,
+      connected: walletConnected,
       getWalletClient,
       signMessageAsync: async (args: { message: string }) => {
         if (
@@ -161,12 +167,9 @@ const useWallet = (): UseWalletReturnType => {
     web3Modal.isConnected,
     mobileSDK.connected,
     mobileSDK.metadata?.name,
-    address,
     walletConnected,
-    isMagic,
     walletConnectInstanceRef,
     coinbaseMobileSDKInstanceRef,
-    web3,
     getWalletClient,
   ]);
