--- packages/app/hooks/use-web3.ts
@@ -1,13 +1,0 @@
-import { useContext } from "react";
-
-import { Web3Context } from "app/context/web3-context";
-
-export function useWeb3() {
-  const context = useContext(Web3Context);
-
-  if (!context) {
-    throw "You need to add `Web3Provider` to your root component";
-  }
-
-  return context;
-}
