+++ packages/app/providers/app-providers.tsx
@@ -21,7 +21,6 @@ import { MuteProvider } from "app/providers/mute-provider";
 import { SWRProvider } from "app/providers/swr-provider";
 import { UserProvider } from "app/providers/user-provider";
 import { WalletMobileSDKProvider } from "app/providers/wallet-mobile-sdk-provider";
-import { Web3Provider } from "app/providers/web3-provider";
 
 import { WalletProvider } from "./wallet-provider";
 
@@ -33,36 +32,32 @@ export const AppProviders = ({ children }: { children: React.ReactNode }) => {
           <ColorSchemeProvider>
             <ReactionProvider>
               <WalletMobileSDKProvider>
-                <Web3Provider>
-                  <WalletProvider>
-                    <AlertProvider>
-                      <LightBoxProvider>
-                        <SnackbarProvider>
-                          <NavigationProvider>
-                            <SWRProvider>
-                              <AuthProvider>
-                                <UserProvider>
-                                  <BottomSheetModalProvider>
-                                    {/* @ts-ignore */}
-                                    <FeedProvider>
-                                      <MuteProvider>
-                                        <ClaimProvider>
-                                          <DropProvider>
-                                            {children}
-                                          </DropProvider>
-                                        </ClaimProvider>
-                                      </MuteProvider>
-                                    </FeedProvider>
-                                  </BottomSheetModalProvider>
-                                </UserProvider>
-                              </AuthProvider>
-                            </SWRProvider>
-                          </NavigationProvider>
-                        </SnackbarProvider>
-                      </LightBoxProvider>
-                    </AlertProvider>
-                  </WalletProvider>
-                </Web3Provider>
+                <WalletProvider>
+                  <AlertProvider>
+                    <LightBoxProvider>
+                      <SnackbarProvider>
+                        <NavigationProvider>
+                          <SWRProvider>
+                            <AuthProvider>
+                              <UserProvider>
+                                <BottomSheetModalProvider>
+                                  {/* @ts-ignore */}
+                                  <FeedProvider>
+                                    <MuteProvider>
+                                      <ClaimProvider>
+                                        <DropProvider>{children}</DropProvider>
+                                      </ClaimProvider>
+                                    </MuteProvider>
+                                  </FeedProvider>
+                                </BottomSheetModalProvider>
+                              </UserProvider>
+                            </AuthProvider>
+                          </SWRProvider>
+                        </NavigationProvider>
+                      </SnackbarProvider>
+                    </LightBoxProvider>
+                  </AlertProvider>
+                </WalletProvider>
               </WalletMobileSDKProvider>
             </ReactionProvider>
           </ColorSchemeProvider>
