+++ packages/app/providers/auth-provider.tsx
@@ -16,7 +16,6 @@ import { AuthContext } from "app/context/auth-context";
 import { useAccessTokenManager } from "app/hooks/auth/use-access-token-manager";
 import { useFetchOnAppForeground } from "app/hooks/use-fetch-on-app-foreground";
 import { useWalletMobileSDK } from "app/hooks/use-wallet-mobile-sdk";
-import { useWeb3 } from "app/hooks/use-web3";
 import * as accessTokenStorage from "app/lib/access-token";
 import { deleteAccessToken, useAccessToken } from "app/lib/access-token";
 import { Analytics, EVENTS } from "app/lib/analytics";
@@ -60,7 +59,6 @@ export function AuthProvider({
   const { mutate } = useSWRConfig();
   const web3Modal = useWeb3Modal();
   const mobileSDK = useWalletMobileSDK();
-  const { setWeb3 } = useWeb3();
   const { magic } = useMagic();
   const { setTokens, refreshTokens } = useAccessTokenManager();
   const fetchOnAppForeground = useFetchOnAppForeground();
@@ -97,19 +95,13 @@ export function AuthProvider({
           user_id: res?.data?.profile?.profile_id,
         });
 
-        if (router.canGoBack()) {
-          router.pop();
-        } else {
-          router.replace("/");
-        }
-
         return res;
       }
 
       setAuthenticationStatus("UNAUTHENTICATED");
       throw "Login failed";
     },
-    [setTokens, setAuthenticationStatus, fetchOnAppForeground, mutate, router]
+    [setTokens, setAuthenticationStatus, fetchOnAppForeground, mutate]
   );
   /**
    * Log out the customer if logged in, and clear auth cache.
@@ -140,7 +132,6 @@ export function AuthProvider({
 
       magic?.user?.logout();
 
-      setWeb3(null);
       setAuthenticationStatus("UNAUTHENTICATED");
       mutate(null);
 
@@ -148,15 +139,7 @@ export function AuthProvider({
 
       router.push("/");
     },
-    [
-      onWagmiDisconnect,
-      web3Modal,
-      mobileSDK,
-      magic?.user,
-      setWeb3,
-      mutate,
-      router,
-    ]
+    [onWagmiDisconnect, web3Modal, mobileSDK, magic?.user, mutate, router]
   );
   const doRefreshToken = useCallback(async () => {
     setAuthenticationStatus("REFRESHING");
