--- packages/app/providers/wallet-provider.web.tsx
@@ -1,3 +1,0 @@
-export const WalletProvider = ({ children }: { children: React.ReactNode }) => {
-  return children;
-};
